fib: use 32 bits per-source reference counter
The 16 bits reference counter limited the number of interfaces
per FIB table to 65K from a given source. Some use cases, for example
GTP-U tunnels require much more interfaces than that.
This change increases the size of the reference counter to 32 bits.

Type: fix

+++ src/vnet/fib/fib_table.h
@@ -83,7 +83,7 @@ typedef struct fib_table_t_
     /**
      * per-source number of locks on the table
      */
-    u16 *ft_locks;
+    u32 *ft_locks;
     u32 ft_total_locks;
 
     /**
